requirements.txt
python-dotenv
openai
accelerate

gradio
gradio_client
scipy
scikit-image
pydantic
huggingface-hub

transformers[agents]
duckduckgo_search
fastapi[standard]

ngrok
uvicorn
fastapi
loguru